requirements.txt
rdflib
flask
SPARQLWrapper
requests
numpy
matplotlib
amadeus
foursquare
pif
python-dotenv
apscheduler